+++ packages/icons/src/components/IconDnsZone.tsx
@@ -1,0 +1,4 @@
+/* auto-generated file */
+import { IconNetwork as Icon } from "@tabler/icons-react";
+import type { FC } from "react";
+export const IconDnsZone = Icon as FC;