+++ packages/icons/src/components/IconEmailArchive.tsx
@@ -1,0 +1,4 @@
+/* auto-generated file */
+import { IconCloudLock as Icon } from "@tabler/icons-react";
+import type { FC } from "react";
+export const IconEmailArchive = Icon as FC;