+++ packages/icons/src/components/IconQuote.tsx
@@ -1,0 +1,4 @@
+/* auto-generated file */
+import { IconBlockquote as Icon } from "@tabler/icons-react";
+import type { FC } from "react";
+export const IconQuote = Icon as FC;