+++ packages/icons/src/iconSet.ts
@@ -1,0 +1,266 @@
+/* auto-generated file */
+import { IconAscending as Ascending } from "./components/IconAscending.tsx";
+import { IconAttachment as Attachment } from "./components/IconAttachment.tsx";
+import { IconBackLink as BackLink } from "./components/IconBackLink.tsx";
+import { IconBold as Bold } from "./components/IconBold.tsx";
+import { IconCheck as Check } from "./components/IconCheck.tsx";
+import { IconCheckboxChecked as CheckboxChecked } from "./components/IconCheckboxChecked.tsx";
+import { IconCheckboxEmpty as CheckboxEmpty } from "./components/IconCheckboxEmpty.tsx";
+import { IconCheckboxIndeterminate as CheckboxIndeterminate } from "./components/IconCheckboxIndeterminate.tsx";
+import { IconChevronDown as ChevronDown } from "./components/IconChevronDown.tsx";
+import { IconChevronLeft as ChevronLeft } from "./components/IconChevronLeft.tsx";
+import { IconChevronRight as ChevronRight } from "./components/IconChevronRight.tsx";
+import { IconChevronUp as ChevronUp } from "./components/IconChevronUp.tsx";
+import { IconClose as Close } from "./components/IconClose.tsx";
+import { IconCode as Code } from "./components/IconCode.tsx";
+import { IconContextMenu as ContextMenu } from "./components/IconContextMenu.tsx";
+import { IconContextMenuVertical as ContextMenuVertical } from "./components/IconContextMenuVertical.tsx";
+import { IconCopy as Copy } from "./components/IconCopy.tsx";
+import { IconDescending as Descending } from "./components/IconDescending.tsx";
+import { IconDelete as Delete } from "./components/IconDelete.tsx";
+import { IconDownload as Download } from "./components/IconDownload.tsx";
+import { IconEdit as Edit } from "./components/IconEdit.tsx";
+import { IconExternalLink as ExternalLink } from "./components/IconExternalLink.tsx";
+import { IconFilter as Filter } from "./components/IconFilter.tsx";
+import { IconForwardLink as ForwardLink } from "./components/IconForwardLink.tsx";
+import { IconHide as Hide } from "./components/IconHide.tsx";
+import { IconHome as Home } from "./components/IconHome.tsx";
+import { IconItalic as Italic } from "./components/IconItalic.tsx";
+import { IconLink as Link } from "./components/IconLink.tsx";
+import { IconLogout as Logout } from "./components/IconLogout.tsx";
+import { IconMaximize as Maximize } from "./components/IconMaximize.tsx";
+import { IconMenu as Menu } from "./components/IconMenu.tsx";
+import { IconMinimize as Minimize } from "./components/IconMinimize.tsx";
+import { IconMinus as Minus } from "./components/IconMinus.tsx";
+import { IconNotification as Notification } from "./components/IconNotification.tsx";
+import { IconOrderedList as OrderedList } from "./components/IconOrderedList.tsx";
+import { IconPause as Pause } from "./components/IconPause.tsx";
+import { IconPlus as Plus } from "./components/IconPlus.tsx";
+import { IconQuote as Quote } from "./components/IconQuote.tsx";
+import { IconRadioOff as RadioOff } from "./components/IconRadioOff.tsx";
+import { IconRadioOn as RadioOn } from "./components/IconRadioOn.tsx";
+import { IconRedo as Redo } from "./components/IconRedo.tsx";
+import { IconRefresh as Refresh } from "./components/IconRefresh.tsx";
+import { IconSave as Save } from "./components/IconSave.tsx";
+import { IconSearch as Search } from "./components/IconSearch.tsx";
+import { IconSend as Send } from "./components/IconSend.tsx";
+import { IconSettings as Settings } from "./components/IconSettings.tsx";
+import { IconShare as Share } from "./components/IconShare.tsx";
+import { IconShow as Show } from "./components/IconShow.tsx";
+import { IconSorting as Sorting } from "./components/IconSorting.tsx";
+import { IconStart as Start } from "./components/IconStart.tsx";
+import { IconStrikeThrough as StrikeThrough } from "./components/IconStrikeThrough.tsx";
+import { IconUndo as Undo } from "./components/IconUndo.tsx";
+import { IconUnorderedList as UnorderedList } from "./components/IconUnorderedList.tsx";
+import { IconUpload as Upload } from "./components/IconUpload.tsx";
+import { IconView as View } from "./components/IconView.tsx";
+import { IconAI as AI } from "./components/IconAI.tsx";
+import { IconAnalytics as Analytics } from "./components/IconAnalytics.tsx";
+import { IconApp as App } from "./components/IconApp.tsx";
+import { IconAutoresponder as Autoresponder } from "./components/IconAutoresponder.tsx";
+import { IconBackup as Backup } from "./components/IconBackup.tsx";
+import { IconCamera as Camera } from "./components/IconCamera.tsx";
+import { IconChangelog as Changelog } from "./components/IconChangelog.tsx";
+import { IconCertificate as Certificate } from "./components/IconCertificate.tsx";
+import { IconContainer as Container } from "./components/IconContainer.tsx";
+import { IconContract as Contract } from "./components/IconContract.tsx";
+import { IconContractPartner as ContractPartner } from "./components/IconContractPartner.tsx";
+import { IconContributor as Contributor } from "./components/IconContributor.tsx";
+import { IconCronjob as Cronjob } from "./components/IconCronjob.tsx";
+import { IconCustomer as Customer } from "./components/IconCustomer.tsx";
+import { IconDashboard as Dashboard } from "./components/IconDashboard.tsx";
+import { IconDatabase as Database } from "./components/IconDatabase.tsx";
+import { IconDate as Date } from "./components/IconDate.tsx";
+import { IconDeliveryBox as DeliveryBox } from "./components/IconDeliveryBox.tsx";
+import { IconDesktop as Desktop } from "./components/IconDesktop.tsx";
+import { IconDirectory as Directory } from "./components/IconDirectory.tsx";
+import { IconDnsSubZone as DnsSubZone } from "./components/IconDnsSubZone.tsx";
+import { IconDnsZone as DnsZone } from "./components/IconDnsZone.tsx";
+import { IconDomain as Domain } from "./components/IconDomain.tsx";
+import { IconEmail as Email } from "./components/IconEmail.tsx";
+import { IconEmailArchive as EmailArchive } from "./components/IconEmailArchive.tsx";
+import { IconExtension as Extension } from "./components/IconExtension.tsx";
+import { IconFeedback as Feedback } from "./components/IconFeedback.tsx";
+import { IconFile as File } from "./components/IconFile.tsx";
+import { IconForwardAddress as ForwardAddress } from "./components/IconForwardAddress.tsx";
+import { IconImage as Image } from "./components/IconImage.tsx";
+import { IconInvite as Invite } from "./components/IconInvite.tsx";
+import { IconInvoice as Invoice } from "./components/IconInvoice.tsx";
+import { IconKnowledgeHint as KnowledgeHint } from "./components/IconKnowledgeHint.tsx";
+import { IconLanguage as Language } from "./components/IconLanguage.tsx";
+import { IconLeadFyndr as LeadFyndr } from "./components/IconLeadFyndr.tsx";
+import { IconLock as Lock } from "./components/IconLock.tsx";
+import { IconLockOpen as LockOpen } from "./components/IconLockOpen.tsx";
+import { IconMainUser as MainUser } from "./components/IconMainUser.tsx";
+import { IconMarketplace as Marketplace } from "./components/IconMarketplace.tsx";
+import { IconMember as Member } from "./components/IconMember.tsx";
+import { IconMittwald as Mittwald } from "./components/IconMittwald.tsx";
+import { IconMobile as Mobile } from "./components/IconMobile.tsx";
+import { IconMonitoring as Monitoring } from "./components/IconMonitoring.tsx";
+import { IconMove as Move } from "./components/IconMove.tsx";
+import { IconNumber as Number } from "./components/IconNumber.tsx";
+import { IconOrder as Order } from "./components/IconOrder.tsx";
+import { IconPassword as Password } from "./components/IconPassword.tsx";
+import { IconPath as Path } from "./components/IconPath.tsx";
+import { IconPayment as Payment } from "./components/IconPayment.tsx";
+import { IconPicture as Picture } from "./components/IconPicture.tsx";
+import { IconProject as Project } from "./components/IconProject.tsx";
+import { IconRegistry as Registry } from "./components/IconRegistry.tsx";
+import { IconSecurity as Security } from "./components/IconSecurity.tsx";
+import { IconServer as Server } from "./components/IconServer.tsx";
+import { IconSsh as Ssh } from "./components/IconSsh.tsx";
+import { IconSshKey as SshKey } from "./components/IconSshKey.tsx";
+import { IconSshSftp as SshSftp } from "./components/IconSshSftp.tsx";
+import { IconStar as Star } from "./components/IconStar.tsx";
+import { IconStarFilled as StarFilled } from "./components/IconStarFilled.tsx";
+import { IconSubdomain as Subdomain } from "./components/IconSubdomain.tsx";
+import { IconSupport as Support } from "./components/IconSupport.tsx";
+import { IconTerminate as Terminate } from "./components/IconTerminate.tsx";
+import { IconTicket as Ticket } from "./components/IconTicket.tsx";
+import { IconTime as Time } from "./components/IconTime.tsx";
+import { IconUser as User } from "./components/IconUser.tsx";
+import { IconVhost as Vhost } from "./components/IconVhost.tsx";
+import { IconVolume as Volume } from "./components/IconVolume.tsx";
+import { IconDanger as Danger } from "./components/IconDanger.tsx";
+import { IconFailed as Failed } from "./components/IconFailed.tsx";
+import { IconInfo as Info } from "./components/IconInfo.tsx";
+import { IconPending as Pending } from "./components/IconPending.tsx";
+import { IconSucceeded as Succeeded } from "./components/IconSucceeded.tsx";
+import { IconSuccess as Success } from "./components/IconSuccess.tsx";
+import { IconUnavailable as Unavailable } from "./components/IconUnavailable.tsx";
+import { IconWarning as Warning } from "./components/IconWarning.tsx";
+
+export const defaultIconSet = {
+  Ascending,
+  Attachment,
+  BackLink,
+  Bold,
+  Check,
+  CheckboxChecked,
+  CheckboxEmpty,
+  CheckboxIndeterminate,
+  ChevronDown,
+  ChevronLeft,
+  ChevronRight,
+  ChevronUp,
+  Close,
+  Code,
+  ContextMenu,
+  ContextMenuVertical,
+  Copy,
+  Descending,
+  Delete,
+  Download,
+  Edit,
+  ExternalLink,
+  Filter,
+  ForwardLink,
+  Hide,
+  Home,
+  Italic,
+  Link,
+  Logout,
+  Maximize,
+  Menu,
+  Minimize,
+  Minus,
+  Notification,
+  OrderedList,
+  Pause,
+  Plus,
+  Quote,
+  RadioOff,
+  RadioOn,
+  Redo,
+  Refresh,
+  Save,
+  Search,
+  Send,
+  Settings,
+  Share,
+  Show,
+  Sorting,
+  Start,
+  StrikeThrough,
+  Undo,
+  UnorderedList,
+  Upload,
+  View,
+  AI,
+  Analytics,
+  App,
+  Autoresponder,
+  Backup,
+  Camera,
+  Changelog,
+  Certificate,
+  Container,
+  Contract,
+  ContractPartner,
+  Contributor,
+  Cronjob,
+  Customer,
+  Dashboard,
+  Database,
+  Date,
+  DeliveryBox,
+  Desktop,
+  Directory,
+  DnsSubZone,
+  DnsZone,
+  Domain,
+  Email,
+  EmailArchive,
+  Extension,
+  Feedback,
+  File,
+  ForwardAddress,
+  Image,
+  Invite,
+  Invoice,
+  KnowledgeHint,
+  Language,
+  LeadFyndr,
+  Lock,
+  LockOpen,
+  MainUser,
+  Marketplace,
+  Member,
+  Mittwald,
+  Mobile,
+  Monitoring,
+  Move,
+  Number,
+  Order,
+  Password,
+  Path,
+  Payment,
+  Picture,
+  Project,
+  Registry,
+  Security,
+  Server,
+  Ssh,
+  SshKey,
+  SshSftp,
+  Star,
+  StarFilled,
+  Subdomain,
+  Support,
+  Terminate,
+  Ticket,
+  Time,
+  User,
+  Vhost,
+  Volume,
+  Danger,
+  Failed,
+  Info,
+  Pending,
+  Succeeded,
+  Success,
+  Unavailable,
+  Warning,
+} as const;
+
+export default defaultIconSet;