+++ packages/remote-react-components/src/tests/visual/TextField.browser.test.tsx
@@ -78,3 +78,34 @@ test.each(testEnvironments)(
     await testScreenshot("TextField - text entered");
   },
 );
+
+test.each(testEnvironments)(
+  "TextField edge cases (%s)",
+  async ({
+    testScreenshot,
+    render,
+    components: { Flex, TextField, Label, Button, IconStar },
+  }) => {
+    await render(
+      <Flex direction="column" gap="m">
+        <TextField value="Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque eius quam quas vel voluptas, ullam aliquid fugit. Voluptate harum accusantium rerum ullam modi blanditiis vitae.">
+          <Label>Default</Label>
+        </TextField>
+        <TextField
+          type="password"
+          value="Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque eius quam quas vel voluptas, ullam aliquid fugit. Voluptate harum accusantium rerum ullam modi blanditiis vitae."
+        >
+          <Label>Password</Label>
+        </TextField>
+        <TextField value="Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque eius quam quas vel voluptas, ullam aliquid fugit. Voluptate harum accusantium rerum ullam modi blanditiis vitae.">
+          <Label>CustomButton</Label>
+          <Button aria-label="Custom">
+            <IconStar />
+          </Button>
+        </TextField>
+      </Flex>,
+    );
+
+    await testScreenshot("TextField edge cases");
+  },
+);