+++ CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.741](https://github.com/mittwald/flow/compare/0.2.0-alpha.740...0.2.0-alpha.741) (2026-03-20)
+
+### Bug Fixes
+
+* **Icons:** fix icons size for Font Awesome ([#2392](https://github.com/mittwald/flow/issues/2392)) ([e2001b7](https://github.com/mittwald/flow/commit/e2001b7caa119db8ce8a11a17b1a998e74a9c131))
+
+# [0.2.0-alpha.740](https://github.com/mittwald/flow/compare/0.2.0-alpha.739...0.2.0-alpha.740) (2026-03-20)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.739](https://github.com/mittwald/flow/compare/0.2.0-alpha.738...0.2.0-alpha.739) (2026-03-20)
+
+### Features
+
+* **List:** add radio filter reset button ([#2376](https://github.com/mittwald/flow/issues/2376)) ([65c012c](https://github.com/mittwald/flow/commit/65c012ccca81ca57a639e660823d84230612d186))
+
+# [0.2.0-alpha.738](https://github.com/mittwald/flow/compare/0.2.0-alpha.737...0.2.0-alpha.738) (2026-03-20)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.737](https://github.com/mittwald/flow/compare/0.2.0-alpha.736...0.2.0-alpha.737) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.736](https://github.com/mittwald/flow/compare/0.2.0-alpha.735...0.2.0-alpha.736) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.735](https://github.com/mittwald/flow/compare/0.2.0-alpha.734...0.2.0-alpha.735) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.734](https://github.com/mittwald/flow/compare/0.2.0-alpha.733...0.2.0-alpha.734) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.733](https://github.com/mittwald/flow/compare/0.2.0-alpha.732...0.2.0-alpha.733) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.732](https://github.com/mittwald/flow/compare/0.2.0-alpha.731...0.2.0-alpha.732) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.731](https://github.com/mittwald/flow/compare/0.2.0-alpha.730...0.2.0-alpha.731) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.730](https://github.com/mittwald/flow/compare/0.2.0-alpha.729...0.2.0-alpha.730) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.729](https://github.com/mittwald/flow/compare/0.2.0-alpha.728...0.2.0-alpha.729) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.728](https://github.com/mittwald/flow/compare/0.2.0-alpha.727...0.2.0-alpha.728) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.727](https://github.com/mittwald/flow/compare/0.2.0-alpha.726...0.2.0-alpha.727) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.726](https://github.com/mittwald/flow/compare/0.2.0-alpha.725...0.2.0-alpha.726) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.725](https://github.com/mittwald/flow/compare/0.2.0-alpha.724...0.2.0-alpha.725) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.724](https://github.com/mittwald/flow/compare/0.2.0-alpha.723...0.2.0-alpha.724) (2026-03-18)
+
+### Bug Fixes
+
+* **TextField:** fix padding ([#2391](https://github.com/mittwald/flow/issues/2391)) ([7d54850](https://github.com/mittwald/flow/commit/7d54850dc6f1a2e43f493f2518d33819bbaedea2))
+
+# [0.2.0-alpha.723](https://github.com/mittwald/flow/compare/0.2.0-alpha.722...0.2.0-alpha.723) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.722](https://github.com/mittwald/flow/compare/0.2.0-alpha.721...0.2.0-alpha.722) (2026-03-18)
+
+### Features
+
+* Add interchangeable icon packages ([61faf2d](https://github.com/mittwald/flow/commit/61faf2db0c65ce1f4e299bcd3d6cfab073d97ca5))
+
+# [0.2.0-alpha.721](https://github.com/mittwald/flow/compare/0.2.0-alpha.720...0.2.0-alpha.721) (2026-03-18)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
 # [0.2.0-alpha.720](https://github.com/mittwald/flow/compare/0.2.0-alpha.719...0.2.0-alpha.720) (2026-03-18)
 
 **Note:** Version bump only for package @mittwald/flow-project